+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/SportsViewModel.java
@@ -4,6 +4,7 @@ import android.app.Application;
 import android.util.Log;
 
 import com.sdzshn3.android.news247.Fragments.SportsNewsFragment;
+import com.sdzshn3.android.news247.Repositories.SportsRepository;
 import com.sdzshn3.android.news247.Retrofit.ApiService;
 import com.sdzshn3.android.news247.Retrofit.Article;
 import com.sdzshn3.android.news247.Retrofit.Client;
@@ -20,54 +21,19 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 public class SportsViewModel extends AndroidViewModel {
-    private static MutableLiveData<List<Article>> data = new MutableLiveData<>();
 
-    private ApiService apiService;
-    private static Call<NewsModel> call;
+    private SportsRepository repository;
 
     public SportsViewModel(@NonNull Application application) {
         super(application);
-        apiService = Client.getApiService();
-        loadData();
+        repository = new SportsRepository();
     }
 
     public LiveData<List<Article>> getData() {
-        return data;
-    }
-
-    private void loadData() {
-        call = apiService.getResponse(SportsNewsFragment.URL);
-        call.enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("SportsViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        return repository.getData();
     }
 
     public void refresh() {
-        call = apiService.getResponse(SportsNewsFragment.URL);
-        call.clone().enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("SportsViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository.refresh();
     }
 }
